pyproject.toml
[project]
name = "dd_pyparse"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "compressed-rtf",
    "ijson",
    "loguru",
    "msg-parser",
    "numpy",
    "opencv-python",
    "pdfminer.six",
    "pillow",
    "pillow-avif-plugin",
    "pillow-heif",
    "python-magic",
    "pydantic",
    "pydantic-settings",
    "python-docx",
    "python-iso639",
    "python-pptx",
    "py7zr",
    "rarfile",
    "tabulate",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "black", "flake8", "isort"]
api = ["uvicorn", "gunicorn", "python-multipart", "aiohttp", "fastapi"]

[tool.setuptools]
package-dir = {"" = "src"}

[project.scripts]
api = "dd_pyparse.interfaces._fastapi:main"